pyproject.toml
[project]
name = "jupiter-pdf-qa-agent"
version = "0.1.0"
description = "A multi-turn PDF QA agent using Streamlit and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.8",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]

[project.optional-dependencies]
dev = [
   "pytest>=8.4.2",
   "ruff>=0.13.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
